Update README.md
Every industry has the potential to benefit from data research and analysis in some way or form. For this first module, our goal is to identify one industry/organization that could see benefit from a research proposal. One of the largest industries that has yet to fully capitalize on the benefits of data analysis is the financial and banking industry. Banking has become almost second-nature in current society with pretty much all forms of saving, spending, and borrowing money being conducted through a bank account. However, there have been many risks identified over the years to the successful long term implementation of analytics in banking. This proposal will delve into the history of analytics within banking and identify the risks that have plagued its success. Following this, a problem will be identified that the use of analytics can address, and the goals of the research. Lastly, I will share how the research proposal will provide benefit to the banking industry.